Moved Dropdown component to FormInputs directory
